refactor(customer): migrate components and types to shared directory
- Updated imports in customer-normal components to use shared types and components.
- Moved constants, info button base, ingredient tab content types,
move button, result card atoms, and tag group to a shared directory.
- Refactored type definitions for better organization and reusability.
- Removed redundant code from customer-rare components by utilizing shared components.
