+++ .github/plans/逻辑解耦/00-总体任务与进度.md
@@ -4,9 +4,9 @@
 
 - 总体状态：进行中
 - 当前阶段：阶段 0 / 阶段 1 / 阶段 2 / 阶段 3 / 阶段 4 / 阶段 5 / 阶段 6 文档已齐备
-- 当前焦点：阶段 2
-- 当前优先级：阶段 2
-- 当前里程碑：阶段 2 已确认 PR-2.1 ~ PR-2.3 满足首版目标，PR-2.4 已完成，下一步推进 PR-2.5
+- 当前焦点：阶段 3
+- 当前优先级：阶段 3
+- 当前里程碑：阶段 2 已完成，阶段 3 的 PR-3.1 已完成，下一步推进 PR-3.2
 - 最后更新：2026-04-27
 - 当前基线验证：`pnpm exec tsc --noEmit` 通过；`pnpm lint` 0 error / 12 warnings；`pnpm build` 通过，静态页 `137/137`
 
@@ -24,8 +24,8 @@
 | ------ | ---------------------------------- | ------------------------------------------------------------------------------------------------------------------ | --------------------- | ------ | ------------------------------------------ |
 | 阶段 0 | 纯函数下沉                         | `buildRecipeSuitabilityRows`、`buildBeverageSuitabilityRows`、`getIngredientScoreChanges` 等                       | 无                    | 已完成 | `PR-0.1 ~ PR-0.6` 已落地，尚未切换旧调用点 |
 | 阶段 1 | 建立稳定的 state / view-model 载体 | store computed、route business hooks、推荐卡片 hook                                                                | 阶段 0 核心纯函数完成 | 已完成 | `PR-1.1 ~ PR-1.5` 已落地                   |
-| 阶段 2 | 视图组件削薄                       | tab content、saved meal、result/customer/info、route content 改读新 view-model                                     | 阶段 1 关键节点完成   | 进行中 | PR-2.1 ~ PR-2.4 已完成或判定无需空改动     |
-| 阶段 3 | 共享子块整合                       | 共享 tab body、shared row / list / avatar 片段                                                                     | 阶段 2 稳定           | 未开始 |                                            |
+| 阶段 2 | 视图组件削薄                       | tab content、saved meal、result/customer/info、route content 改读新 view-model                                     | 阶段 1 关键节点完成   | 已完成 | PR-2.1 ~ PR-2.7 已完成或判定无需空改动     |
+| 阶段 3 | 共享子块整合                       | 共享 tab body、shared row / list / avatar 片段                                                                     | 阶段 2 稳定           | 进行中 | PR-3.1 已完成，下一步 PR-3.2               |
 | 阶段 4 | 领域单例升级                       | `Recipe` / `CustomerRare` / `CustomerNormal` 中层聚合方法                                                          | 阶段 0-2 经验沉淀     | 未开始 |                                            |
 | 阶段 5 | 清理与防退化                       | 死代码删除、缓存替换、repo memory 更新                                                                             | 前面阶段稳定          | 未开始 |                                            |
 | 阶段 6 | 阶段 5 之后的补遗                  | store factory、Tab 视图共享化、resultCard/savedMealCard/customerCard 小共享、`useCustomerRouteData` 提升、低优收尾 | 未开始（非主路径）    |        |
@@ -38,8 +38,8 @@
 | [00-总体任务与进度.md](00-总体任务与进度.md)                   | 总控        | 总任务、步骤、进度、风险锚点                               | 当前文件           |
 | [01-阶段0-纯函数下沉.md](01-阶段0-纯函数下沉.md)               | 阶段 0 细化 | PR-0.1 ~ PR-0.6、输入输出、调用点、落地顺序                | 进行中             |
 | [02-阶段1-state与view-model.md](02-阶段1-state与view-model.md) | 阶段 1 细化 | store computed、business hooks、推荐卡片 view-model        | 进行中             |
-| [03-阶段2-视图削薄.md](03-阶段2-视图削薄.md)                   | 阶段 2 细化 | 各页面组件切换策略、依赖切换与回归点                       | 进行中             |
-| [04-阶段3-共享整合.md](04-阶段3-共享整合.md)                   | 阶段 3 细化 | 共享子块、shared surface 与禁止合并边界                    | 未开始             |
+| [03-阶段2-视图削薄.md](03-阶段2-视图削薄.md)                   | 阶段 2 细化 | 各页面组件切换策略、依赖切换与回归点                       | 已完成             |
+| [04-阶段3-共享整合.md](04-阶段3-共享整合.md)                   | 阶段 3 细化 | 共享子块、shared surface 与禁止合并边界                    | 进行中             |
 | [05-阶段4-领域升级.md](05-阶段4-领域升级.md)                   | 阶段 4 细化 | 领域收口、单例方法边界与禁止上提清单                       | 未开始             |
 | [06-阶段5-清理与防退化.md](06-阶段5-清理与防退化.md)           | 阶段 5 细化 | ESLint、缓存、死代码、repo memory 与最终验证               | 未开始             |
 | [07-阶段6-补遗.md](07-阶段6-补遗.md)                           | 阶段 6 细化 | store factory、Tab 视图共享化、小共享、提升 hook、低优收尾 | 未开始（非主路径） |
@@ -49,4 +49,5 @@
 
 1. 阶段 0 已完成：`app/utils/customer/shared/` 已落地 `PR-0.1 ~ PR-0.6` 的纯函数与 shared barrel，且保持旧页面/store 调用点不变。
 2. 阶段 1 已完成：`customer-rare` / `customer-normal` store 已承接表格 rows、`ingredientScoreChanges` 与 `savedCustomerMealsWithEvaluation`，稀客侧已补齐 `currentMealPrice`、`unsatisfiedSelectionTip`、`bondRewards`、`useSuggestedMealsViewModel`，route content 也已改读 `useCustomerRouteData`、`useIngredientRouteData` 与 `useDocumentTitle`。
-3. 阶段 2 正在推进：PR-2.1 ~ PR-2.3 当前仓库已满足首版目标，无需空改动；PR-2.4 已完成 rare `suggestedMealCard` 的业务输入拼装下沉，下一步推进 PR-2.5。
+3. 阶段 2 已完成：PR-2.1 ~ PR-2.3、PR-2.5 ~ PR-2.7 当前仓库已满足首版目标，PR-2.4 完成了 rare `suggestedMealCard` 的业务输入拼装下沉。
+4. 阶段 3 已启动：PR-3.1 已将 `InfoButtonBase`、`TagGroup`、`Plus`、`UnknownItem`、`MoveButton`、tab 常量与基础类型 surface 收口到 `app/(pages)/customer-shared/`，并清除了 `customer-normal` 对 `customer-rare` 的页面级横向 import；下一步推进 PR-3.2。
