+++ .github/plans/逻辑解耦/03-阶段2-视图削薄.md
@@ -14,10 +14,10 @@
 
 ## 当前状态
 
-- 状态：进行中
-- 已确认满足首版目标：PR-2.1、PR-2.2、PR-2.3
+- 状态：已完成
+- 已确认满足首版目标：PR-2.1、PR-2.2、PR-2.3、PR-2.5、PR-2.6、PR-2.7
 - 已完成：PR-2.4
-- 下一步：PR-2.5
+- 下一步：阶段 3 / PR-3.1
 - 当前基线验证：`pnpm exec tsc --noEmit` 通过；`pnpm lint` 0 error / 12 warnings（现有 `onClick` deprecation warnings，非本轮新增）；`pnpm build` 通过，静态页 `137/137`
 
 ## 推荐顺序
@@ -312,6 +312,12 @@
 - 当前套餐价格与显示名在料理、酒水、额外食材变化后持续更新。
 - normal 仍只要求料理即可给出评级 / 保存提示，不意外引入酒水或点单标签作为必需条件。
 
+### 当前判断
+
+- rare `customerCard.tsx` 已消费 `unsatisfiedSelectionTip.rating`，rare `resultCard.tsx` 已消费 `currentMealPrice` 与 `unsatisfiedSelectionTip.save`，评级提示与保存提示不再各自手工拼装。
+- `currentMealDisplayMeta` 当前仍未在代码中落地，本阶段不为清理少量展示分支反向新增 store surface，避免把 PR-2.5 做成大改。
+- normal 继续保留静态提示文案，符合“只做必要清理、不强行补复杂 selection tip 或 display meta”的边界。
+
 ## PR-2.6：InfoButton
 
 ### 来源文件
@@ -353,6 +359,11 @@
 - 默认展开项与旧逻辑一致。
 - normal 的 description / chat / rating / help 四节保持不变。
 - 打开 Info 后仍写入 `info` query 参数，关闭后仍移除该参数；携带 `info` 参数直达页面时仍能自动打开对应弹窗。
+
+### 当前判断
+
+- rare `infoButton.tsx` 已消费 `bondRewards`，剩余的 `hasBondRewards` 判断与 `defaultExpandedKeys` 都是贴近 UI 的本地决策，不再属于阶段 2 需要移走的业务聚合。
+- normal `infoButton.tsx` 继续保留 description / chat / rating / help 的页级结构与默认展开逻辑，符合计划中的差异保留要求。
 
 ## PR-2.7：Route [[...paths]] / content.tsx
 
@@ -404,6 +415,11 @@
 - 食材过滤仍按带流行趋势修正后的标签工作。
 - 食材 tab 只在已选料理后可用。
 - rare route 仍在 ResultCard 与 SavedMealCard 之间渲染 `SuggestedMealCard`；normal route 不会多出该卡片。
+
+### 当前判断
+
+- rare / normal `[[...paths]]/content.tsx` 已分别消费 `useCustomerRouteData`、`useIngredientRouteData` 与 `useDocumentTitle`，route 级顾客列表、食材列表与标题副作用流水线已经收口。
+- 当前 content 层剩余的是 `validateName`、`router.replace`、Tabs、侧边按钮、mounted guard 与 rare 插槽布局，符合本阶段保留 route 壳的约束。
 
 ## 可继续保持不对称的部分
