+++ .github/plans/逻辑解耦/04-阶段3-共享整合.md
@@ -12,6 +12,13 @@
 - 不把仍在变化的 view-model 直接塞进 shared 组件。
 - 不把稀客专属语义写成默认分支，再让普客去兼容它。
 
+## 当前状态
+
+- 状态：进行中
+- 已完成：PR-3.1
+- 下一步：PR-3.2
+- 当前基线验证：`pnpm exec tsc --noEmit` 通过；`pnpm lint` 0 error / 12 warnings（现有 `onClick` deprecation warnings，非本轮新增）；`pnpm build` 通过，静态页 `137/137`
+
 ## 共享原则
 
 ### 可以共享的前提
@@ -88,18 +95,17 @@
 - `app/(pages)/customer-rare/constants.tsx`
 - 对应 normal 页面对这些 rare 文件的横向复用入口
 
-### 当前横向 import 盘点（实施前必须完整覆盖）
+### 本轮覆盖的横向 import
 
 - `app/(pages)/customer-normal/layout.tsx` -> `@/(pages)/customer-rare/layout`
 - `app/(pages)/customer-normal/infoButton.tsx` -> `@/(pages)/customer-rare/infoButtonBase`
 - `app/(pages)/customer-normal/constants.tsx` -> `@/(pages)/customer-rare/constants`
-- `app/(pages)/customer-normal/beverageTabContent.tsx` -> `@/(pages)/customer-rare/beverageTabContent`
-- `app/(pages)/customer-normal/recipeTabContent.tsx` -> `@/(pages)/customer-rare/recipeTabContent`
 - `app/(pages)/customer-normal/resultCard.tsx` -> `@/(pages)/customer-rare/resultCard`
 - `app/(pages)/customer-normal/savedMealCard.tsx` -> `@/(pages)/customer-rare/savedMealCard`
 - `app/(pages)/customer-normal/tagGroup.tsx` -> `@/(pages)/customer-rare/tagGroup`
 - `app/(pages)/customer-normal/types.d.ts` -> `@/(pages)/customer-rare/types`
 - `app/(pages)/customer-normal/ingredientTabContent.tsx` -> `@/(pages)/customer-rare/ingredientTabContent`
+- `app/components/customerRareTutorial.tsx` -> `@/(pages)/customer-rare/infoButtonBase`
 
 ### 要做什么
 
@@ -126,10 +132,16 @@
 - 行为与视觉完全不变。
 - normal -> rare 的横向 import 清单应在迁移后清零或仅保留经明确批准的过渡项。
 
-### 状态说明（当前未实施）
+### 本轮落地记录
 
-- 本节以下内容均为阶段 3 目标态，当前代码仍以 page-local 复用为主。
-- shared 目录与 barrel 的具体落点、导入切换顺序和清理项需在实施 PR 中逐条验收后再回写为“已完成”。
+- 新增 `app/(pages)/customer-shared/` 下的 `infoButtonBase.tsx`、`tagGroup.tsx`、`resultCardAtoms.tsx`、`moveButton.tsx`、`constants.tsx`、`types.d.ts` 与 `ingredientTabContentTypes.d.ts`，作为 PR-3.1 的 shared surface。
+- `customer-normal` 页面对 `infoButtonBase`、`tagGroup`、`resultCard` 原子、`savedMealCard` 按钮原子、`constants`、`types`、`ingredientTabContent` props 与 `layout` 的 rare 页面级直连已全部切除。
+- rare 侧保留 `infoButtonBase.tsx`、`tagGroup.tsx`、`constants.tsx` 等本地兼容导出壳，避免一次性重写全部调用路径。
+
+### 当前判断
+
+- 当前 `app/(pages)/customer-normal/**/*.ts*` 中已不存在 `@/(pages)/customer-rare/...` 的页面级横向 import；PR-3.1 目标态已经达成。
+- 后续阶段 3 应转向 table shell、ingredient presenter 等更高层共享块，而不是继续在 page-local 兼容壳上做无收益重排。
 
 ## PR-3.2：共享 table shell
