+++ app/(pages)/customer-normal/tagGroup.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/(pages)/customer-rare/tagGroup';
+export { default } from '@/(pages)/customer-shared/tagGroup';
