+++ app/(pages)/customer-rare/resultCard.tsx
@@ -1,14 +1,10 @@
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import useBreakpoint from 'use-breakpoint';
 import { useVibrate } from '@/hooks';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-	faCircleXmark,
-	faPlus,
-	faQuestion,
-} from '@fortawesome/free-solid-svg-icons';
+import { faCircleXmark } from '@fortawesome/free-solid-svg-icons';
 
 import {
 	Button,
@@ -19,6 +15,7 @@ import {
 	cn,
 } from '@/design/ui/components';
 
+import { Plus, UnknownItem } from '@/(pages)/customer-shared/resultCardAtoms';
 import Placeholder from '@/components/placeholder';
 import Price from '@/components/price';
 import Sprite from '@/components/sprite';
@@ -31,60 +28,7 @@ import {
 import { customerRareStore as customerStore, globalStore } from '@/stores';
 import { checkA11yConfirmKey, toArray } from '@/utilities';
 
-interface IPlusProps extends Pick<HTMLSpanElementAttributes, 'className'> {
-	size?: number;
-}
-
-export const Plus = memo<IPlusProps>(function Plus({ className, size = 1 }) {
-	const remString = `${size}rem`;
-
-	return (
-		<span
-			className={cn('mx-1 text-center leading-none', className)}
-			style={{ fontSize: remString, width: remString }}
-		>
-			<FontAwesomeIcon icon={faPlus} />
-		</span>
-	);
-});
-
-interface IUnknownItemProps extends Pick<
-	HTMLSpanElementAttributes,
-	'className' | 'title'
-> {
-	size?: number;
-}
-
-export const UnknownItem = memo<IUnknownItemProps>(function UnknownItem({
-	className,
-	size = 2,
-	title,
-}) {
-	const remString = `${size}rem`;
-
-	return (
-		<Tooltip showArrow content={title} offset={7 + -8 * (size - 2)}>
-			<span
-				role="img"
-				title={title}
-				className={cn(
-					'outline-3 flex items-center justify-center rounded-small p-0.5 text-center leading-none outline-double',
-					className
-				)}
-				style={{
-					fontSize: remString,
-					height: remString,
-					width: remString,
-				}}
-			>
-				<FontAwesomeIcon
-					icon={faQuestion}
-					className="!h-full rotate-12"
-				/>
-			</span>
-		</Tooltip>
-	);
-});
+export { Plus, UnknownItem } from '@/(pages)/customer-shared/resultCardAtoms';
 
 function IngredientsList() {
 	const vibrate = useVibrate();
