+++ app/(pages)/customer-rare/savedMealCard.tsx
@@ -1,13 +1,8 @@
-import { Fragment, memo } from 'react';
+import { Fragment } from 'react';
 
 import { usePictureInPicture, useVibrate, useViewInNewWindow } from '@/hooks';
 
 import { Divider } from '@heroui/divider';
-import {
-	FontAwesomeIcon,
-	type FontAwesomeIconProps,
-} from '@fortawesome/react-fontawesome';
-import { faArrowDown, faArrowUp } from '@fortawesome/free-solid-svg-icons';
 
 import {
 	Avatar,
@@ -22,6 +17,10 @@ import {
 	cn,
 } from '@/design/ui/components';
 
+import {
+	type IMoveButtonProps,
+	MoveButton,
+} from '@/(pages)/customer-shared/moveButton';
 import { Plus } from './resultCard';
 import TagGroup from './tagGroup';
 import { trackEvent } from '@/components/analytics';
@@ -38,64 +37,10 @@ import {
 import { customerRareStore as customerStore, globalStore } from '@/stores';
 import { checkLengthEmpty, copyArray } from '@/utilities';
 
-const moveButtonDirectionMap = { down: 0, up: 1 } as const;
-
-type TMoveButtonDirection = ExtractCollectionValue<
-	typeof moveButtonDirectionMap
->;
-
-export interface IMoveButtonProps extends Pick<
-	FontAwesomeIconProps,
-	'onClick'
-> {
-	direction: TMoveButtonDirection;
-	isDisabled: boolean;
-}
-
-const MoveButtonComponent = memo<IMoveButtonProps>(function MoveButton({
-	direction,
-	isDisabled,
-	onClick,
-}) {
-	return (
-		<Tooltip
-			showArrow
-			content={
-				direction === moveButtonDirectionMap.down
-					? isDisabled
-						? '已是末项'
-						: '下移此项'
-					: isDisabled
-						? '已是首项'
-						: '上移此项'
-			}
-			offset={5}
-			placement="left"
-			size="sm"
-		>
-			<FontAwesomeIcon
-				icon={
-					direction === moveButtonDirectionMap.down
-						? faArrowDown
-						: faArrowUp
-				}
-				size="1x"
-				onClick={onClick}
-				role="button"
-				className={cn(
-					'cursor-pointer text-default transition-colors hover:text-default-400 motion-reduce:transition-none',
-					{ 'cursor-not-allowed hover:text-default-200': isDisabled }
-				)}
-			/>
-		</Tooltip>
-	);
-});
-
-export const MoveButton = MoveButtonComponent as typeof MoveButtonComponent & {
-	direction: typeof moveButtonDirectionMap;
-};
-
-MoveButton.direction = moveButtonDirectionMap;
+export {
+	type IMoveButtonProps,
+	MoveButton,
+} from '@/(pages)/customer-shared/moveButton';
 
 export default function SavedMealCard() {
 	const {
