+++ app/(pages)/customer-shared/types.d.ts
@@ -1,0 +1,40 @@
+import { type SortDescriptor } from '@heroui/table';
+
+export type TTabVisibilityState = 'collapse' | 'expand';
+
+export interface ICustomerTabStyle {
+	ariaLabel: string;
+	buttonNode: ReactNodeWithoutBoolean;
+	classNames: { content: string; sideButtonGroup: string };
+}
+
+export type TCustomerTabStyleMap = Record<
+	TTabVisibilityState,
+	ICustomerTabStyle
+>;
+
+export interface IIngredientsTabStyle {
+	ariaLabel: string;
+	buttonNode: ReactNodeWithoutBoolean;
+	classNames: { content: string; sideButtonGroup: string };
+}
+
+export type TIngredientsTabStyleMap = Record<
+	TTabVisibilityState,
+	IIngredientsTabStyle
+>;
+
+export interface ITableColumn<T extends string> {
+	key: T;
+	label: string;
+	sortable: boolean;
+}
+
+export interface ITableSortDescriptor<T extends string> extends SortDescriptor {
+	column?: T;
+	direction?: SortDescriptor['direction'];
+	lastColumn?: T;
+	time?: number;
+}
+
+export type TTab = 'beverage' | 'customer' | 'ingredient' | 'recipe';
